+++ src/read.c
@@ -123,9 +123,9 @@ doread()
 					: card_msgs[scroll->o_id % (SIZE(card_msgs)-1)]);
 		}
 		/* Make a credit card number */
-		pline("\"%d0%d %d%d1 0%d%d0\"", ((scroll->o_id % 89)+10), (scroll->o_id % 4),
-				(((scroll->o_id * 499) % 899999) + 100000), (scroll->o_id % 10),
-				(!(scroll->o_id % 3)), ((scroll->o_id * 7) % 10));
+		pline("\"%d0%d %d%d1 0%d%d0\"", (int)((scroll->o_id % 89)+10), (int)(scroll->o_id % 4),
+				(int)(((scroll->o_id * 499) % 899999) + 100000), (int)(scroll->o_id % 10),
+				(int)(!(scroll->o_id % 3)), (int)((scroll->o_id * 7) % 10));
 		u.uconduct.literate++;
 		return 1;
 #endif	/* TOURIST */
